many changes
* Removed various duplicate and unused methods.
* MainPage now opens directly if you previously logged in.
