+++ OneDrive-Cloud-Player/OneDrive-Cloud-Player.csproj
@@ -127,6 +127,7 @@
     <Compile Include="Services\Helpers\GraphHelper.cs" />
     <Compile Include="Services\Helpers\GraphAuthHelper.cs" />
     <Compile Include="Services\Utilities\JsonHandler.cs" />
+    <Compile Include="Services\Utilities\WindowSwitcher.cs" />
     <Compile Include="ViewModels\LoginPageViewModel.cs" />
     <Compile Include="ViewModels\MainPageViewModel.cs" />
     <Compile Include="ViewModels\StartPageViewModel.cs" />
