+++ OneDrive-Cloud-Player/Services/Utilities/WindowSwitcher.cs
@@ -1,0 +1,31 @@
+﻿using System;
+using System.Threading.Tasks;
+using Windows.ApplicationModel.Core;
+using Windows.UI.Core;
+using Windows.UI.ViewManagement;
+using Windows.UI.Xaml;
+using Windows.UI.Xaml.Controls;
+
+namespace OneDrive_Cloud_Player.Services.Utilities
+{
+	public static class WindowSwitcher
+	{
+       public static async Task<bool> TryOpenNewWindow(Type page)
+       {
+            CoreApplicationView newView = CoreApplication.CreateNewView();
+            int newViewId = 0;
+            await newView.Dispatcher.RunAsync(CoreDispatcherPriority.High, () =>
+            {
+                Frame frame = new Frame();
+                frame.Navigate(page);
+                Window.Current.Content = frame;
+                // You have to activate the window in order to show it later.
+                Window.Current.Activate();
+
+                newViewId = ApplicationView.GetForCurrentView().Id;
+            });
+            bool viewShown = await ApplicationViewSwitcher.TryShowAsStandaloneAsync(newViewId);
+            return viewShown;
+        }
+    }
+}
